requirements.txt
asgiref==3.4.1
cachetools==5.2.0
debugpy==1.6.0
decorator==5.1.1
defusedxml==0.7.1
distlib==0.3.4
dj-database-url==0.5.0
dj-email-url==1.0.5
Django==3.2.6
django-bootstrap4==2.2.0
factory-boy==3.2.1
gitdb==4.0.9
GitPython==3.1.27
gunicorn==20.1.0
idna==3.3
importlib-metadata==4.11.4
iso3166==2.0.2
itsdangerous==2.1.2
jedi==0.18.1
Jinja2==3.1.2
MarkupSafe==2.1.1
matplotlib==3.5.2
matplotlib-inline==0.1.3
numpy==1.21.6
oauthlib==3.2.0
pandas==1.3.5
pipenv==2022.9.2
psutil==5.9.1
psycopg2==2.9.4
psycopg2-binary==2.9.3
pyarrow==8.0.0
PyAutoGUI==0.9.53
pycparser==2.21
pydeck==0.7.1
PyGetWindow==0.0.9
Pygments==2.12.0
pyparsing==3.0.9
pyperclip==1.8.2
PyScreeze==0.1.28
python-dateutil==2.8.2
python-stdnum==1.17
pytweening==1.0.4
pytz==2021.1
soupsieve==2.3.1
speckenv==2.0
sqlparse==0.4.1
streamlit==1.10.0
tzdata==2022.1
tzlocal==4.2
Unipath==1.1
urllib3==1.26.9
validators==0.19.0
virtualenv==20.14.1
webencodings==0.5.1
Werkzeug==2.1.2
python-dotenv==0.21.0
python-telegram-bot==20.1
schedule==1.1.0